requirements.txt
# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Asterisk
ari==0.1.3
requests==2.31.0

# Transcription
faster-whisper==1.2.0
soundfile==0.13.1
scipy==1.15.3
numba==0.62.1

# Sentiment Analysis
textblob==0.17.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Utilities
python-dotenv==1.0.0

# WebSocket client (for ARI)
websocket-client==1.7.0